Add files via upload
This is my first commit where I have two versions of the same script. The TestScript_01 is a very generic one with my own assumptions based on the first meeting on Jan 31st. 

The TestScript_02.m is a later version with the updates and changes mentioned in the 2nd meeting on Feb 5th. 

Currently, I am working on the Data_Preprocessing_function.m file to make the data preprocessing a piece of cake. The function takes in the inputs and generates outputs that facilitate Parameter Estimation. None of the matlab built-in data structures are altered. 

For example, 
1. The groupedData object is the required data format to store data needed for fitting using fitproblem , sbiofit , and sbiofitmixed . It is a table -like object with a few differences. The groupedData object has two additional properties to identify the independent variable and an optional grouping variable.

2. The estimatedInfo object contains information about estimated model quantities (species, parameters, or compartments). Use this object to specify which quantities in a SimBiology® model are estimated, what parameter transforms are used, and optionally, the initial estimates for these quantities when using sbiofit or sbiofitmixed.

3. Doses: In addition to manually constructing dose objects using sbiodose, if the input groupedData object has dosing information, you can use the createDoses method to construct doses.